refactor sidebar path
