+++ apps/website/app/app/layout/Sidebar/Item/Logout.tsx
@@ -1,5 +1,5 @@
 import { ArrowLeftOnRectangleIcon } from "@heroicons/react/20/solid";
-import { Menu, MenuItem } from "@/app/components/Sidebar";
+import { Menu, MenuItem } from "@/components/Sidebar";
 import { Suspense } from "react";
 import { CSRFTokenField } from "@/auth/components";
