+++ apps/website/app/app/layout/Sidebar/Item/SongRequest.tsx
@@ -1,5 +1,5 @@
 import { MusicalNoteIcon } from "@heroicons/react/20/solid";
-import { MenuItem } from "@/app/components/Sidebar";
+import { MenuItem } from "@/components/Sidebar";
 
 const MusicIcon = <MusicalNoteIcon className="h-4 w-4" />;
