+++ apps/website/app/app/layout/Sidebar/index.tsx
@@ -6,7 +6,7 @@ import {
   SidebarContent,
   SidebarHeader,
   SidebarFooter,
-} from "@/app/components/Sidebar";
+} from "@/components/Sidebar";
 import { UserIcon } from "@heroicons/react/20/solid";
 import Image from "next/image";
 import { SongRequestItem, ToggleCollapsedMenu } from "./Item";
