+++ apps/website/app/app/profile/page.tsx
@@ -3,6 +3,11 @@ import { PageHeader } from "@station/client/layout";
 import { getSession } from "@/auth/server";
 import { getCSRFToken } from "@/auth/csrf";
 import ProfileForm from "./Form";
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "ข้อมูลส่วนตัว",
+};
 
 export default async function Profile() {
   const csrfToken = getCSRFToken(headers());
