+++ apps/website/components/Sidebar/useDefaultOpen.ts
